refactor: inline mixin logic into vaadin-tab
